Fix several warnings, detecetd by GCC
- m_curCharId should be quint64 type;
- unhandled enum value and missing return (!!!!)
- missing base class initializer in derived's copy constructor

+++ qmlevemon_app.h
@@ -63,7 +63,7 @@ protected:
     QQuickWindow *m_mainWindow;
     PortraitCache *m_portraitCache;
     PeriodicalRefresher *m_refresher;
-    int m_curCharId;
+    quint64 m_curCharId;
 };
